requirements.txt
Flask>=3.0.0
Werkzeug>=3.0.0
pymongo[srv]>=4.0.0
python-dotenv>=1.0.0
PyJWT>=2.0.0
google-generativeai>=0.5.0